Update DigiBruteDroid.ino
Fixed the todo list. It should work with 5 digits now. What I did was to set a boolean with the default (false). When a == 10, it will set the boolean to true. With the if conditions to testfor what keystrokes should be added, one tests for whether the bool is false and one test for whether it is true. If false, only send keystrokes of int variables a, b, c, and d. If true, send keystrokes of all 5 integers.